Separated math cmd group into cmds directory, added loading code in main file. math 그룹의 커맨드들을 cmds 폴더로 옮기고, main 파일에 불러오는 코드를 적음.
#15
