+++ main.py
@@ -24,6 +24,13 @@ def run():
     async def on_ready():                                           # When bot is online.
         logger.info(f"User: {bot.user} (ID: {bot.user.id})")        # Show message with the format I set.
 
+        # from cmds directory, get every files which name finish with .py
+        for cmd_file in settings.CMDS_DIR.glob("*.py"):
+            # ignore init file.
+            if cmd_file.name != "__init__.py":
+                # last 3 characters are ".py"
+                await bot.load_extension(f"cmds.{cmd_file.name[:-3]}")
+
     @bot.event
     async def on_command_error(ctx, error):                         # Global error handler.
         if isinstance(error, commands.MissingRequiredArgument):
@@ -53,24 +60,6 @@ def run():
     @bot.command()
     async def choices(ctx, *options): 
         await ctx.send(random.choice(options))              # Return one random element from tokens.
-
-    @bot.group()                                            # Make the group of commands.
-    async def math(ctx):                                    # Function name is going to be the group name.
-        if ctx.invoked_subcommand is None:                  # Error handling for unmatched sub_command.
-            await ctx.send(f"No, {ctx.subcommand_passed} does not belong to math")
-
-    @math.command()                                         # bot.command() -> math.command() # Group the command.
-    async def add(ctx, one : int, two : int):               # Set the type of input value.
-        await ctx.send(one + two)                           # Return added result of two input values.
-
-    @add.error                                              # Set error handler for the command "!add".
-    async def add_error(ctx, error):
-        if isinstance(error, commands.MissingRequiredArgument):     # If argument is missing.
-            await ctx.send("handled error locally")
-            
-    @math.command()
-    async def substract(ctx, one : int, two : int):
-        await ctx.send(one - two)
 
     @bot.command()
     async def say3(ctx, what = "WHAT?", why = "WHY?"):      # Take two words as token for what and why.
